convert int to long int

+++ Mini-max sum.c
@@ -7,10 +7,10 @@
 int main()
 {
 
-    int arr[5],max=0,min,sum=0;
+    long int arr[5],max=0,min,sum=0;
     for(int i=0; i<5; i++)
     {
-        scanf("%d",&arr[i]);
+        scanf("%ld",&arr[i]);
         sum+=arr[i];
     }
     min =arr[0];
@@ -23,5 +23,5 @@ int main()
         if(arr[i]<min)
             min=arr[i];
     }
-    printf("%d %d",sum-max,sum-min);
+    printf("%ld %ld",sum-max,sum-min);
 }
